change type

+++ lib.cpp
@@ -1,7 +1,6 @@
 #include <vowpalwabbit/vw.h>
 #include "parse_example_json.h"
 #include "lib.hpp"
-#include <iostream>
 
 
 #define HANDLE_VW_ERRORS                                           \
@@ -115,7 +114,7 @@ VW_DLL_MEMBER void VW_CALLING_CONV VW_MultiLinePredictSafe(VW_HANDLE handle, VW_
     END_HANDLE_VW_ERRORS(error,)
 }
 
-VW_DLL_MEMBER uint32_t VW_CALLING_CONV VW_GetAction(VW_EXAMPLE e, size_t i) { 
+VW_DLL_MEMBER size_t VW_CALLING_CONV VW_GetAction(VW_EXAMPLE e, size_t i) { 
   example * ex = static_cast<example*>(e);
   return ex->pred.a_s[i].action;
 }
