+++ lib.hpp
@@ -1,4 +1,5 @@
 
+#include <stdint.h>
 #include <vowpalwabbit/vwdll.h>
 
 #ifdef __cplusplus
@@ -14,7 +15,7 @@ extern "C"
     VW_DLL_MEMBER VW_EXAMPLE VW_CALLING_CONV VW_ReadJSONExampleSafe(VW_HANDLE handle, const char* line, size_t* example_count, VW_ERROR *error);
     VW_DLL_MEMBER void VW_CALLING_CONV VW_MultiLineLearnSafe(VW_HANDLE handle, VW_EXAMPLE* example_handles, size_t example_count, VW_ERROR *error);
     VW_DLL_MEMBER void VW_CALLING_CONV VW_MultiLinePredictSafe(VW_HANDLE handle, VW_EXAMPLE* example_handles, size_t example_count, VW_ERROR *error);
-    VW_DLL_MEMBER uint32_t VW_CALLING_CONV VW_GetAction(VW_EXAMPLE e, size_t i);
+    VW_DLL_MEMBER size_t VW_CALLING_CONV VW_GetAction(VW_EXAMPLE e, size_t i);
 
 #ifdef __cplusplus
 }
